1.build_dataset，資料來源與產生training dataset邏輯分開以便使用 2.multithread版本產生訓練資料(build_dataset) 3.train改用前幾週資料預測本週資料

+++ build_dataset.py
@@ -9,68 +9,73 @@
 import pandas as pd
 from datetime import datetime
 import datasource.mongodb as mongodb
-from sklearn.model_selection import train_test_split
 import numpy as np
-import matplotlib.pyplot as plt
-
 dateFormat = r"%Y-%m-%dT%H:%M:%S%z"
-debug = False
 
 
 def parse_park_name(park):
     return park['CarParkName']['Zh_tw']
 
 
-"""
-資料庫取得所需資料
-"""
-# 停車場ID, 星期幾, 從0點開始累積得分鐘數, 時間, 剩餘車位數
-parkFrame = pd.DataFrame(
-    columns=["CarParkID", "dayOfWeek",
-             "culMinutes", "datetime", "AvailableSpaces"]
-)
-parkCol = mongodb.obtain_ParkingAvailability()
-# 取哪個欄位
-# projection = {'_id': 1, 'ParkingAvailabilities.$': 1}
-if debug == False:
-    cursor = parkCol.find()
-else:
-    cursor = parkCol.find().skip(1000).limit(100)
+def build_DataFrame(parkingDataCursor):
+    """
+    MongoDB每筆ParkingAvailability(Cursor)轉為DataFrame格式
+    """
+    # 停車場ID, 停車場名稱, 資料收集時間, 總車位數, 剩餘車位數, 星期幾, 從0點開始累積得分鐘數,
+    parkFrame = pd.DataFrame(
+        columns=["CarParkID", "CarParkName", "DataCollectTime",
+                 "NumberOfSpaces", "AvailableSpaces", "dayOfWeek", "culMinutes"]
+    )
 
-dataIdx = 0
-for cur in cursor:
-    parkList = cur['ParkingAvailabilities']
-    # 這時間點蒐集的所有停車場資料
-    collectDatetime = datetime.strptime(cur['SrcUpdateTime'], dateFormat)
-    # 從0點開始累積得分鐘數
-    culMinutes = collectDatetime.hour * 60 + collectDatetime.minute
-    for park in parkList:
-        # 1: 營業, 0:未營業
-        if park['ServiceStatus'] == '0':
-            print(parse_park_name(park) + " 未營業")
-            continue
+    # 用index新增dataFrame一筆資料
+    dataIdx = 0
+    for cur in parkingDataCursor:
+        parkList = cur['ParkingAvailabilities']
+        # 這時間點蒐集的所有停車場資料
+        collectDatetime = datetime.strptime(cur['SrcUpdateTime'], dateFormat)
+        # 從0點開始累積得分鐘數
+        culMinutes = collectDatetime.hour * 60 + collectDatetime.minute
+        for park in parkList:
+            # 跳過未營業停車場
+            # 1: 營業, 0:未營業
+            if park['ServiceStatus'] == 0:
+                print(parse_park_name(park) + " 未營業")
+                continue
 
-        data = {
-            "CarParkID": park['CarParkID'],
-            "datetime": collectDatetime,
-            "culMinutes": culMinutes,
-            "AvailableSpaces": park['AvailableSpaces']
-        }
-        parkFrame.loc[dataIdx] = data
-        dataIdx += 1
-        if dataIdx % 600 == 0:
-            print(f'處理至第{dataIdx}筆')
-# 轉換這日期是星期幾
-parkFrame["dayOfWeek"] = parkFrame["datetime"].dt.dayofweek
-# parkFrame.hist(bins=50, figsize=(20,15))
-# plt.show()
+            # 只取SpaceType:1(自小客車)的資料
+            availTypeList = park['Availabilities']
+            hasCarPark = False
+            for avail in availTypeList:
+                if avail['SpaceType'] == 1:
+                    hasCarPark = True
+                    numberOfSpaces = avail['NumberOfSpaces']
+                    availableSpaces = avail['AvailableSpaces']
+            # 此停車場沒有自小客車車位
+            if hasCarPark == False:
+                print(
+                    f"停車場: {park['CarParkID']}/{parse_park_name(park)} 未設置小客車車位")
+                continue
 
-"""
-建立dataset
-data和target兩組array，用index對應
-"""
-data = parkFrame[['CarParkID', 'dayOfWeek', 'culMinutes']].to_numpy()
-target = parkFrame['AvailableSpaces'].to_numpy()
+            data = {
+                "CarParkID": park['CarParkID'],
+                "CarParkName": parse_park_name(park),
+                "DataCollectTime": collectDatetime,
+                "NumberOfSpaces": numberOfSpaces,
+                "AvailableSpaces": availableSpaces,
+                "culMinutes": culMinutes,
+            }
+            parkFrame.loc[dataIdx] = data
+            dataIdx += 1
+            if dataIdx % 600 == 0:
+                print(f'處理至第{dataIdx}筆')
+
+    if parkFrame.empty:
+        # 沒資料代表都沒營業或都沒小客車停車場，這不太可能，肯定程式哪邊出錯
+        raise ValueError("DataFrame為空")
+
+    # 轉換這日期是星期幾
+    parkFrame["dayOfWeek"] = parkFrame["DataCollectTime"].dt.dayofweek
+    return parkFrame
 
 # TODO 測試用，看資料分布
 # res = pd.Series(parkFrame['CarParkID']).value_counts(normalize=True)
@@ -80,31 +85,71 @@ target = parkFrame['AvailableSpaces'].to_numpy()
 # r = pd.Series(X_train[:, 0]).value_counts(normalize=True)
 # print(r)
 
-# TODO 看是要寫成json、csv還是存入MongoDB
 
-
-def save_to_db(parkFrame: pd.DataFrame):
+def save_to_db(parkFrame: pd.DataFrame, clean_col=False):
     db = mongodb.obtain_db()
     dataCol = db['Dataset']
-    # 清除舊有資料
-    dataCol.delete_many({})
+    if clean_col:
+        # 清除舊有資料
+        dataCol.delete_many({})
+
     # records: 多個物件，每個物件都有key和value
-    insert = parkFrame[['CarParkID', 'dayOfWeek',
-                        'culMinutes', 'AvailableSpaces']].to_dict('records')
+    # 指定欄位
+    # insert = parkFrame[['CarParkID', 'dayOfWeek',
+    #                     'culMinutes', 'AvailableSpaces']].to_dict('records')
+    # 全部欄位
+    # TODO 因為timestamp輸入MongoDB會產生時區消失導致時間錯誤的問題，所以先用字串存日期
+    copy_frame = parkFrame.copy()
+    copy_frame['DataCollectTime'] = copy_frame['DataCollectTime'].dt.strftime(
+        dateFormat)
+    insert = copy_frame.to_dict('records')
     dataCol.insert_many(insert)
 
 
-def save_to_numpyFile(data: np.ndarray, target: np.ndarray):
-    park_filename = 'park_dict.npz'
+def save_to_numpyFile(park_filename: str, data: np.ndarray, target: np.ndarray):
     np.savez(park_filename, data=data, target=target)
     # 只是載回來確認存入是否正常
     dic = np.load(park_filename, allow_pickle=True)
     X = dic['data']
     y = dic['target']
-    dic.close()
+    # TODO 印出來看測試用
+    # print(X)
     # print(y)
+    dic.close()
 
 
-save_to_numpyFile(data, target)
-save_to_db(parkFrame)
-print('done')
+def trans_to_trainData(parkFrame):
+    """
+    data和target兩組array，用index對應
+    data為停車場ID, 星期幾, 從0點開始累積得分鐘數
+    target(label)為剩餘車位數
+    """
+    data = parkFrame[['CarParkID', 'dayOfWeek', 'culMinutes']].to_numpy()
+    target = parkFrame['AvailableSpaces'].to_numpy()
+    return data, target
+
+
+if __name__ == '__main__':
+    debug = True
+    saveNumpy = False
+    saveDB = True
+
+    parkCol = mongodb.obtain_ParkingAvailability()
+    # 取哪個欄位
+    # projection = {'_id': 1, 'ParkingAvailabilities.$': 1}
+    if debug == False:
+        cursor = parkCol.find()
+    else:
+        # cursor = parkCol.find().skip(1000).limit(2)
+        cursor = parkCol.find().limit(1)
+
+    parkFrame = build_DataFrame(cursor)
+    if saveDB:
+        save_to_db(parkFrame, clean_col=True)
+
+    # 存numpy檔
+    if saveNumpy:
+        data, target = trans_to_trainData(parkFrame)
+        park_numpy_filename = 'park_dict.npz'
+        save_to_numpyFile(park_numpy_filename, data, target)
+    print('done')
